requirements.txt
psycopg2-binary==2.9.9
requests==2.31.0
python-dotenv==1.0.0
Flask==2.3.3
Flask-CORS==4.0.0